REFACTOR #12: DTO 클래스 RequestDTO, ResponseDTO 클래스로 분리
