+++ src/main/java/com/umc/FestieBE/domain/together/api/TogetherController.java
@@ -1,9 +1,10 @@
 package com.umc.FestieBE.domain.together.api;
 
 import com.umc.FestieBE.domain.together.application.TogetherService;
-import com.umc.FestieBE.domain.together.dto.TogetherDTO;
+import com.umc.FestieBE.domain.together.dto.TogetherRequestDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
@@ -17,7 +18,7 @@ public class TogetherController {
     private final TogetherService togetherService;
 
     @PostMapping("/together")
-    public ResponseEntity<Void> createTogether(@Valid @RequestBody TogetherDTO.TogetherRequest request) {
+    public ResponseEntity<Void> createTogether(@Valid @RequestBody TogetherRequestDTO request) {
         togetherService.createTogether(request);
         return ResponseEntity.ok().build();
     }
