+++ src/main/java/com/umc/FestieBE/domain/together/application/TogetherService.java
@@ -3,19 +3,16 @@ package com.umc.FestieBE.domain.together.application;
 import com.umc.FestieBE.domain.festival.dao.FestivalRepository;
 import com.umc.FestieBE.domain.festival.domain.Festival;
 import com.umc.FestieBE.domain.temporary_user.TemporaryUser;
-import com.umc.FestieBE.domain.temporary_user.TemporaryUserRepository;
 import com.umc.FestieBE.domain.temporary_user.TemporaryUserService;
 import com.umc.FestieBE.domain.together.dao.TogetherRepository;
 import com.umc.FestieBE.domain.together.domain.Together;
-import com.umc.FestieBE.domain.together.dto.TogetherDTO;
+import com.umc.FestieBE.domain.together.dto.TogetherRequestDTO;
 import com.umc.FestieBE.global.exception.CustomErrorCode;
 import com.umc.FestieBE.global.exception.CustomException;
 import com.umc.FestieBE.global.type.FestivalType;
 import com.umc.FestieBE.global.type.RegionType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.Arrays;
 
 @Service
 @RequiredArgsConstructor
@@ -26,7 +23,7 @@ public class TogetherService {
 
     private final TemporaryUserService temporaryUserService;
 
-    public void createTogether(TogetherDTO.TogetherRequest request) {
+    public void createTogether(TogetherRequestDTO request) {
         // 임시 유저
         TemporaryUser tempUser = temporaryUserService.createTemporaryUser();
