+++ src/main/java/com/umc/FestieBE/domain/together/dto/TogetherResponseDTO.java
@@ -1,0 +1,8 @@
+package com.umc.FestieBE.domain.together.dto;
+
+import lombok.Getter;
+
+@Getter
+public class TogetherResponseDTO {
+
+}
